Allow reply email with a rewritten to address.
Calvin College in MI is known to rewrite the from/reply-to address of
our email. When a recipient hits Reply, the message is delivered to the
hostname of our server rather than the custom subdomain of the site.

This workaround should allow replies to rewritten to addresses, as long
as the proper message id is present.