Update sync.rb
If there is only a single group in a sync, the sync fails. Wrapping the hashes to force an array to keep from getting the "can't convert string to integer" error that happens when an array index isn't present in the case of a single-record hash. Wrapping other hashed lists (families, people) that are handled the same way.